cheri: tools: perf: Fix integer-to-pointer cast warnings
Explicitly cast through uintptr_t/intptr_t to satisfy the CHERI
compiler  that the integer is being intentionally stored as a
capability  pointer. This eliminates the build warnings and
ensures compatibility with CHERI systems.

+++ tools/perf/bench/breakpoint.c
@@ -7,6 +7,7 @@
 #include <sys/syscall.h>
 #include <sys/ioctl.h>
 #include <sys/time.h>
+#include <inttypes.h>
 #include <pthread.h>
 #include <stddef.h>
 #include <stdlib.h>
@@ -54,7 +55,7 @@ static int breakpoint_setup(void *addr)
 	attr.inherit = 1;
 	attr.exclude_kernel = 1;
 	attr.exclude_hv = 1;
-	attr.bp_addr = (unsigned long)addr;
+	attr.bp_addr = (unsigned long)(uintptr_t)addr;
 	attr.bp_type = HW_BREAKPOINT_RW;
 	attr.bp_len = HW_BREAKPOINT_LEN_1;
 	fd = syscall(SYS_perf_event_open, &attr, 0, -1, -1, 0);
